+++ packages/ack/lib/src/schemas/list_schema.dart
@@ -56,7 +56,7 @@ final class ListSchema<ItemBoundary extends Object, ItemRuntime extends Object>
         name: '$i',
         schema: itemSchema,
         value: item,
-        pathSegment: '$i',
+        pathSegment: i,
       );
       final r = parse
           ? itemSchema.parseWithContext(item, itemCtx)
@@ -122,7 +122,7 @@ final class ListSchema<ItemBoundary extends Object, ItemRuntime extends Object>
         name: '$i',
         schema: itemSchema,
         value: item,
-        pathSegment: '$i',
+        pathSegment: i,
         operation: SchemaOperation.encode,
       );
       try {